requirements.txt
flask
flask-cors
openai
gunicorn
pillow
